实现 VerifySsoToken 接口
